chore!: store brillig memory addresses as `u32` instead of `usize`

+++ acvm-repo/acir/codegen/acir.cpp
@@ -786,7 +786,7 @@ namespace Acir {
     struct MemoryAddress {
 
         struct Direct {
-            uint64_t value;
+            uint32_t value;
 
             friend bool operator==(const Direct&, const Direct&);
             std::vector<uint8_t> bincodeSerialize() const;
@@ -805,7 +805,7 @@ namespace Acir {
         };
 
         struct Relative {
-            uint64_t value;
+            uint32_t value;
 
             friend bool operator==(const Relative&, const Relative&);
             std::vector<uint8_t> bincodeSerialize() const;
@@ -907,7 +907,7 @@ namespace Acir {
 
     struct HeapArray {
         Acir::MemoryAddress pointer;
-        uint64_t size;
+        uint32_t size;
 
         friend bool operator==(const HeapArray&, const HeapArray&);
         std::vector<uint8_t> bincodeSerialize() const;
@@ -1611,7 +1611,7 @@ namespace Acir {
 
         struct Array {
             std::vector<Acir::HeapValueType> value_types;
-            uint64_t size;
+            uint32_t size;
 
             friend bool operator==(const Array&, const Array&);
             std::vector<uint8_t> bincodeSerialize() const;
